+++ gezgin-core/src/commonMain/kotlin/dev/gezgin/core/compose/BottomSheetScene.kt
@@ -1,4 +1,7 @@
-@file:OptIn(ExperimentalMaterial3Api::class)
+@file:OptIn(
+    ExperimentalMaterial3Api::class,
+    dev.gezgin.core.ExperimentalGezginMigrationApi::class,
+)
 
 package dev.gezgin.core.compose