+++ gezgin-core/src/commonMain/kotlin/dev/gezgin/core/compose/EntryAdapter.kt
@@ -1,4 +1,4 @@
-@file:OptIn(GezginInternalApi::class)
+@file:OptIn(GezginInternalApi::class, dev.gezgin.core.ExperimentalGezginMigrationApi::class)
 
 package dev.gezgin.core.compose